Extract ChatService focused modules
