+++ docs/spec-testing.md
@@ -19,6 +19,8 @@ Test HTTP servers that are contacted through `127.0.0.1` bind explicitly to `127
 | `test/workspaceContext.service.test.ts` | Workspace Context service coverage: enabling creates `workspace-context/WORKSPACE_CONTEXT.md`, `context/overview.md`, `references/`, `assets/`, `runs/`, and `state.json`; installs and removes the managed `AGENTS.md` block without overwriting unrelated instructions; resolves global vs workspace processor settings; runs autonomous scan prompts with tool access in the real workspace; runs maintenance over existing context markdown, reference files, asset inventory, and accepted Memory inbox entries; consumes accepted active Memory inbox entries and deletes them; fails visibly when required Memory action output is missing; records run state for initial, scheduled, manual, reset, archive, maintenance, and skipped sources; verifies frequent scans do not starve maintenance; supports stop/clear/repair behavior; lists and reads only markdown files under `workspace-context/context/`; lists/reads/writes/deletes markdown/text references under `workspace-context/references/`; lists/stores/serves/deletes allowlisted non-executable assets under `workspace-context/assets/`; and never creates graph, candidate, MCP, or review artifacts. |
 | `test/chat.workspaceContext.test.ts` | Workspace Context REST coverage for settings, enablement, manual scan, maintenance, stop, clear, instruction repair, context file listing/reads, reference listing/reads/writes/deletes, asset listing/upload/preview/download/delete, conversation-scoped Workspace Context context/reference/asset previews/download validation, disabled/unknown workspace responses, conversation `workspaceContext` hydration, and workspace-scoped `workspace_context_update` emissions. |
 | `test/chat.memory.test.ts` Memory v2 REST coverage | Workspace memory route coverage: `GET /workspaces/:workspaceId/memory/search` rejects empty queries, returns `{ enabled:false, query, results: [] }` while memory is disabled, searches enabled workspace memory, defaults to active/redacted entries, and honors type/status filters including explicit superseded recall; consolidation propose/apply and draft/apply-draft routes cover safe supersession and reviewed rewrites; deleted CLI-capture entries stay hidden after future native captures; removed Memory Review routes return `410`; and conversation responses no longer expose `memoryReview`. |
+| `test/chat.memorySearch.test.ts` | Direct Workspace Memory lexical-search coverage for tokenization, lifecycle/type filters, exact-match and type-aware ranking boosts, BM25-style ordering, recency tie-breaks, and snippet extraction. |
+| `test/chat.extractedModules.test.ts` | Direct coverage for ChatService extraction modules: memory metadata normalization, conversation policy defaults, transcript markdown rendering, session transition persistence, Workspace Context status normalization, KB state snapshot counters, workspace memory lifecycle delegation, legacy migrations, and worktree-isolation bulk toggle orchestration. |
 | `test/frontendRoutes.test.ts` Memory v2 static coverage | Workspace Memory panel static guards assert the frontend exposes search, restore, delete, clear, the search input, type/lifecycle filter controls, global Settings Memory processor status/profile copy, and CSS hooks for search controls, lifecycle status pills, and Memory processor status. The same guards assert Memory Review API helpers, schedule/run UI, dedicated page, composer notification, and dead consolidation review UI are absent. |
 | `test/frontendRoutes.test.ts` Workspace Context static coverage | Static guards assert `AgentApi.workspace` exposes Workspace Context settings, enablement, manual scan, maintenance, stop, clear, repair, context file list/read helpers, reference list/read/save/delete helpers, and asset list/upload/view/download/delete helpers; global Settings exposes the Workspace Context tab with processor defaults, scan/maintenance intervals, and scan/maintenance concurrency; Workspace Settings exposes Overview, Processor, Context, References, Assets, Runs, and Danger Zone sections; run state polling keeps the on-screen Workspace Context tab current, renders runs latest-first by start time, renders run summaries as Markdown, and exposes older/newer run-log paging; context previews are read-only markdown; references and assets share the same two-pane file-browser design as context while keeping reference edit controls and asset upload controls in the preview pane; asset previews handle safe text/markdown/image files; desktop chat resolves Workspace Context context/reference/asset links into file-preview descriptors; old graph/entity/candidate/review endpoints and UI hooks are absent; the composer exposes a Workspace Context notification for running/failed runs; and Workspace Context CSS contains the shared browser plus markdown/reference/asset/run controls without old graph or review classes. |
 | `test/routines.service.test.ts` | Workspace Routines service coverage: disabled workspace no-side-effect list/run behavior, generated authoring files and managed `AGENTS.md` block, proposal marker validation inside the workspace routines folder, schedule validation, install/edit/disable/delete lifecycle, one-shot run folder/input/final output handling, `notify.md` notification routing including global Telegram bot token fallback with workspace chat destinations, process-local Telegram destination pairing through `getUpdates`, scheduler due checks over routines-enabled workspaces, and Telegram request shape. |
